Fixed Classes to be testable, Added Duplicative Record ID Validation and Fixed Concrete Report Strategies Classes

+++ src/main/java/com/movieproject/contexts/FileHandler.java
@@ -10,12 +10,10 @@ import java.io.IOException;
 public class FileHandler {
     private String filePath;
     private String tempFilePath;
-    private String tempDirPath;
 
     public FileHandler(String filePath) {
         this.filePath = filePath;
         this.tempFilePath = "./data/temp_data.csv";
-        this.tempDirPath = "./data/temp";
     }
 
     public String getTempFilePath() {
@@ -26,27 +24,52 @@ public class FileHandler {
         return this.filePath;
     }
 
-    public void performOperation(FileOperation operation) {
+    public boolean performOperation(FileOperation operation) {
         File mainFile = new File(this.filePath);
         new File(this.tempFilePath);
 
         try {
-            operation.execute(mainFile);
-            if (operation instanceof FileUpdateOperation || operation instanceof FileDeleteOperation)
-                this.replaceMainFile();
-        } catch (IOException var5) {
-            IOException e = var5;
-            e.printStackTrace();
+            boolean success = operation.execute(mainFile);
+            if (success && (operation instanceof FileUpdateOperation || operation instanceof FileDeleteOperation)) {
+                return replaceMainFile();
+            }
+            return success;
+        } catch (IOException e) {
+            System.out.println("Error occurred during file operation: " + e.getMessage());
+            deleteTempFile(); // Ensure temp file is deleted on error
+            return false;
         }
 
     }
 
-    public void replaceMainFile() {
+    private boolean replaceMainFile() {
         File tempFile = new File(this.tempFilePath);
         File mainFile = new File(this.filePath);
-        if (mainFile.delete()) {
-            tempFile.renameTo(mainFile);
+
+        if (!tempFile.exists()) {
+            System.out.println("Temp file does not exist.");
+            return false;
         }
 
+        if (mainFile.delete()) {
+            if (tempFile.renameTo(mainFile)) {
+                return true;
+            } else {
+                System.out.println("Failed to rename temp file to main file.");
+                return false;
+            }
+        } else {
+            System.out.println("Failed to delete the original file.");
+            return false;
+        }
+    }
+
+    private void deleteTempFile() {
+        File tempFile = new File(this.tempFilePath);
+        if (tempFile.exists()) {
+            if (!tempFile.delete()) {
+                System.out.println("Failed to delete temp file.");
+            }
+        }
     }
 }
